feat: Replace ncurses with raylib
This commit replaces the ncurses-based terminal UI with a graphical UI powered by the raylib library.

Changes include:
- Added raylib as a dependency using FetchContent.
- Created a new `src/raylib_io.cpp` file to handle all raylib-related I/O and rendering.
- Updated the main game loop to use raylib's window and event handling.
- Removed the old ncurses-based code and dependencies.
